synalm_lmax should be an integer

+++ pysm3/models/dust_realization.py
@@ -135,7 +135,7 @@ class ModifiedBlackBodyRealization(ModifiedBlackBody):
             seeds = (None, None, None)
 
         if synalm_lmax is None:
-            synalm_lmax = min(16384, 2.5 * self.nside)
+            synalm_lmax = int(min(16384, 2.5 * self.nside))
 
         np.random.seed(seeds[0])
